refactor: Baseball 추상화 제거
기존 Baseball과 다른 자식 Baseball에 비해 달라지는 점이 있을 때 추상화를 해야 함.
제네릭으로 List에 들어갈 Type만 바뀌면 되고, 추가적으로 Type에 따라 달라지는 건 없으므로 제거.
