--- src/main/java/baseball/domain/baseball/IntegerBaseball.java
@@ -1,9 +1,0 @@
-package baseball.domain.baseball;
-
-import java.util.List;
-
-public class IntegerBaseball extends Baseball<Integer> {
-
-        //기존 Baseball과 다른 자식 Baseball에 비해 달라지는 점이 있을 때 따로 class로 만드는 것
-
-}
