--- src/main/java/baseball/domain/baseball/StringBaseball.java
@@ -1,8 +1,0 @@
-package baseball.domain.baseball;
-
-import java.util.List;
-
-public class StringBaseball extends Baseball<String> {
-
-
-}
